feat: re-structured language command with sub commands
- Changed command name from `pull-languages` to `languages` across the codebase for consistency and add sub command 'pull'
- Updated resolved path to `.storyblok/languages/spaceId/filename.suffix.json` by default
- Updated launch configuration to reflect the new command name.
- Introduced new actions and constants for managing languages, including fetching and saving language data.
- Added tests for the new languages command and its actions.
- Removed obsolete `pull-languages` command and its associated tests.
