+++ src/commands/languages/index.ts
@@ -2,25 +2,30 @@ import { colorPalette, commands } from '../../constants'
 import { CommandError, handleError, konsola } from '../../utils'
 import { getProgram } from '../../program'
 import { session } from '../../session'
-import { pullLanguages, saveLanguagesToFile } from './actions'
+import { fetchLanguages, saveLanguagesToFile } from './actions'
 import chalk from 'chalk'
 import type { PullLanguagesOptions } from './constants'
 
 const program = getProgram() // Get the shared singleton instance
 
-export const pullLanguagesCommand = program
-  .command('pull-languages')
-  .description(`Download your space's languages schema as json`)
+export const languagesCommand = program
+  .command('languages')
+  .description(`Manage your space's languages`)
   .option('-s, --space <space>', 'space ID')
   .option('-p, --path <path>', 'path to save the file. Default is .storyblok/languages')
+
+languagesCommand
+  .command('pull')
+  .description(`Download your space's languages schema as json`)
   .option('-f, --filename <filename>', 'filename to save the file as <filename>.<suffix>.json')
   .option('--su, --suffix <suffix>', 'suffix to add to the file name (e.g. languages.<suffix>.json). By default, the space ID is used.')
   .action(async (options: PullLanguagesOptions) => {
-    konsola.title(` ${commands.PULL_LANGUAGES} `, colorPalette.PULL_LANGUAGES, 'Pulling languages...')
+    konsola.title(` ${commands.LANGUAGES} `, colorPalette.LANGUAGES, 'Pulling languages...')
     // Global options
     const verbose = program.opts().verbose
     // Command options
-    const { space, path, filename = 'languages', suffix = options.space } = options
+    const { space, path } = languagesCommand.opts()
+    const { filename = 'languages', suffix = options.space } = options
 
     const { state, initializeSession } = session()
     await initializeSession()
@@ -35,14 +40,19 @@ export const pullLanguagesCommand = program
     }
 
     try {
-      const internationalization = await pullLanguages(space, state.password, state.region)
+      const internationalization = await fetchLanguages(space, state.password, state.region)
 
       if (!internationalization || internationalization.languages?.length === 0) {
         konsola.warn(`No languages found in the space ${space}`)
         return
       }
-      await saveLanguagesToFile(space, internationalization, options)
-      konsola.ok(`Languages schema downloaded successfully at ${chalk.hex(colorPalette.PRIMARY)(path ? `${path}/${filename}.${suffix}.json` : `.storyblok/languages/${filename}.${suffix}.json`)}`)
+      await saveLanguagesToFile(space, internationalization, {
+        ...options,
+        path,
+      })
+      const fileName = suffix ? `${filename}.${suffix}.json` : `${filename}.json`
+      const filePath = path ? `${path}/${fileName}` : `.storyblok/languages/${space}/${fileName}`
+      konsola.ok(`Languages schema downloaded successfully at ${chalk.hex(colorPalette.PRIMARY)(filePath)}`)
     }
     catch (error) {
       handleError(error as Error, verbose)
